refactor(opensearch): extract _lib.py and hash full cfg into Job name
