+++ opensearch/chart/files/snapshot-lifecycle/install.py
@@ -24,87 +24,17 @@ the chart (the CRD covers replicaCount + delete actions). Only the
 which the shipped CRD does not yet support
 (opensearch-k8s-operator#1402).
 """
-import base64
-import json
 import os
-import ssl
 import sys
 import time
 from pathlib import Path
-from urllib import error, request
 
-CLUSTER = os.environ["CLUSTER_HOST"].rstrip("/")
-ADMIN_USER = os.environ["ADMIN_USER"]
-ADMIN_PASSWORD = os.environ["ADMIN_PASSWORD"]
+# /scripts is on sys.path automatically because that's where this file lives.
+from _lib import CLUSTER, http, log
+
 STREAMS = os.environ["STREAMS"].split()
 REPOSITORIES = os.environ["REPOSITORIES"].split()
 SCRIPTS = Path("/scripts")
-
-# TLS_SKIP_VERIFY=true disables certificate verification (default: verify).
-# CA_BUNDLE points at a mounted CA bundle file when verifying against a
-# private CA. The chart mounts the cluster's HTTP TLS secret by default.
-SKIP_VERIFY = os.environ.get("TLS_SKIP_VERIFY", "").lower() in ("1", "true", "yes")
-CA_BUNDLE = os.environ.get("CA_BUNDLE") or None
-
-if SKIP_VERIFY:
-    SSL_CTX = ssl.create_default_context()
-    SSL_CTX.check_hostname = False
-    SSL_CTX.verify_mode = ssl.CERT_NONE
-elif CA_BUNDLE:
-    SSL_CTX = ssl.create_default_context(cafile=CA_BUNDLE)
-else:
-    SSL_CTX = ssl.create_default_context()
-
-_AUTH_HEADER = "Basic " + base64.b64encode(
-    f"{ADMIN_USER}:{ADMIN_PASSWORD}".encode()
-).decode()
-
-
-def log(level: str, msg: str, **fields) -> None:
-    """Emit a logfmt line: ts=... level=... msg="..." key=value ..."""
-    parts = [
-        f"ts={time.strftime('%Y-%m-%dT%H:%M:%SZ', time.gmtime())}",
-        f"level={level}",
-        f'msg="{msg}"',
-    ]
-    for k, v in fields.items():
-        s = str(v)
-        if any(c in s for c in ' "\\='):
-            s = '"' + s.replace('\\', '\\\\').replace('"', '\\"') + '"'
-        parts.append(f"{k}={s}")
-    print(" ".join(parts), flush=True)
-
-
-class Response:
-    __slots__ = ("status", "body")
-
-    def __init__(self, status: int, body: bytes):
-        self.status = status
-        self.body = body
-
-    @property
-    def ok(self) -> bool:
-        return 200 <= self.status < 300
-
-    @property
-    def text(self) -> str:
-        return self.body.decode("utf-8", errors="replace")
-
-    def json(self) -> dict:
-        return json.loads(self.body) if self.body else {}
-
-
-def http(method: str, path: str, body: dict | None = None, timeout: int = 30) -> Response:
-    data = json.dumps(body).encode() if body is not None else None
-    req = request.Request(f"{CLUSTER}{path}", data=data, method=method)
-    req.add_header("Authorization", _AUTH_HEADER)
-    if data is not None:
-        req.add_header("Content-Type", "application/json")
-    try:
-        with request.urlopen(req, timeout=timeout, context=SSL_CTX) as resp:
-            return Response(resp.status, resp.read())
-    except error.HTTPError as e:
-        return Response(e.code, e.read())
 
 
 def wait_for_cluster(timeout_s: int = 240) -> None:
@@ -190,6 +120,7 @@ def put_sm_policy(path: str, body: dict) -> None:
 
 def load(filename: str, substitutions: dict | None = None) -> dict:
     """Load a JSON file, applying string substitutions before parsing."""
+    import json
     raw = (SCRIPTS / filename).read_text()
     for key, value in (substitutions or {}).items():
         raw = raw.replace(key, value)
